py-utils/requirements.txt
pysaucenao
python-dotenv
